+++ contracts/token/ERC1155/IERC1155MetadataURI.sol
@@ -9,5 +9,11 @@ import "./IERC1155.sol";
  * in the https://eips.ethereum.org/EIPS/eip-1155#metadata-extensions[EIP].
  */
 interface IERC1155MetadataURI is IERC1155 {
+    /**
+     * @dev Returns the URI for token type `id`.
+     *
+     * If the `\{id\}` substring is present in the URI, it must be replaced by
+     * clients with the actual token type ID.
+     */
     function uri(uint256 id) external view returns (string memory);
 }