+++ contracts/token/ERC721/ERC721Burnable.sol
@@ -11,8 +11,11 @@ import "./ERC721.sol";
  */
 abstract contract ERC721Burnable is Context, ERC721 {
     /**
-     * @dev Burns a specific ERC721 token.
-     * @param tokenId uint256 id of the ERC721 token to be burned.
+     * @dev Burns `tokenId`. See {ERC721-_burn}.
+     *
+     * Requirements:
+     *
+     * - The caller must own `tokenId` or be an approved operator.
      */
     function burn(uint256 tokenId) public virtual {
         //solhint-disable-next-line max-line-length